+++ config-templates/tables.sql
@@ -12,17 +12,17 @@ CREATE TABLE statistics (
     INDEX (year,month),
     INDEX (year,month,day),
     PRIMARY KEY (year, month, day, sourceIdp, service)
-);
+) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 --Tables for mapping identifier to name
 CREATE TABLE identityProvidersMap(
 	entityId VARCHAR(255) NOT NULL,
 	name VARCHAR(255) NOT NULL,
 	PRIMARY KEY (entityId)
-);
+) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 
 CREATE TABLE serviceProvidersMap(
 	identifier VARCHAR(255) NOT NULL,
 	name VARCHAR(255) NOT NULL,
 	PRIMARY KEY (identifier)
-);
+) ENGINE=InnoDB DEFAULT CHARSET=utf8;